+++ src/index.css
@@ -20,4 +20,5 @@ code {
 
 #root {
   background-color: #EEF3F4;
+  position: relative;
 }
